rule_engine_with_ast/requirements.txt
Flask==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
